backend/requirements.txt
alembic==1.4.2
click==7.1.2
ecdsa==0.16.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pycryptodome==3.3.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose-cryptodome==1.3.2
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
